Fixed imports in core modules. Fixed references in all files (hopefully). Fixed test runner paths.
