+++ README.md
@@ -1,0 +1,15 @@
+This code is for controlling the rover by sending data back and forth from an Xbee
+
+To run the Base Station a virtual environment will need to be created.
+
+1) To create the virtual enviroment, you will first need to install the package
+"""pip install virtualenv"""
+
+2) To create a virtual environment. 
+"""virtualenv venv"""
+
+3) To enter the virtual environment call.
+"""source venv/Scripts/activate"""
+
+4) Next install the requirements to run the program
+"""pip install -r requirements.txt"""
