+++ requirements.txt
@@ -9,5 +9,9 @@ binho-host-adapter==0.1.6
 pyftdi==0.56.0
 pyserial==3.5
 pyusb==1.3.1
-sysv_ipc==1.1.0
 typing_extensions==4.12.2
+pytest
+pyyaml
+pygame
+lark
+digi-xbee
