+++ xbee/__main__.py
@@ -1,0 +1,3 @@
+from core.Xbee_refactored import main
+
+main()
