+++ xbee/core/heartbeat.py
@@ -7,7 +7,7 @@ import time
 import struct
 from typing import Optional
 
-from ..CommandCodes import CONSTANTS
+from .CommandCodes import CONSTANTS
 
 class HeartbeatManager:
     """
