--- xbee/tests/__init__.py
@@ -1,3 +1,0 @@
-"""
-Test modules for XBee system validation.
-"""
